project_requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.2
amqp==5.1.1
anyio==3.6.2
apache-airflow==2.5.2
apache-airflow-providers-amazon==8.2.0
apache-airflow-providers-celery==3.1.0
apache-airflow-providers-common-sql==1.3.4
apache-airflow-providers-ftp==3.3.1
apache-airflow-providers-http==4.2.0
apache-airflow-providers-imap==3.1.1
apache-airflow-providers-jdbc==3.3.0
apache-airflow-providers-postgres==5.4.0
apache-airflow-providers-sqlite==3.3.1
apispec==3.3.2
argcomplete==2.1.1
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.12.1
beautifulsoup4==4.12.2
billiard==3.6.4.0
blinker==1.5
boto3==1.26.125
botocore==1.29.125
cachelib==0.9.0
cattrs==22.2.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
commonmark==0.9.1
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.2.35
croniter==1.3.8
cryptography==39.0.2
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.1.1
dnspython==2.3.0
docutils==0.19
email-validator==1.3.1
exceptiongroup==1.1.1
Flask==2.2.3
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
flower==1.2.0
frozenlist==1.3.3
graphviz==0.20.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
humanize==4.6.0
idna==3.4
importlib-metadata==4.13.0
importlib-resources==5.12.0
inflection==0.5.1
itsdangerous==2.1.2
JayDeBeApi==1.2.3
Jinja2==3.1.2
jmespath==1.0.1
JPype1==1.4.1
jsonpath-ng==1.5.3
jsonschema==4.17.3
kombu==5.2.4
lazy-object-proxy==1.9.0
linkify-it-py==2.0.0
lockfile==0.12.2
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.3.5
mdurl==0.1.2
multidict==6.0.4
mypy-boto3-appflow==1.26.162
mypy-boto3-rds==1.26.163
mypy-boto3-redshift-data==1.26.109
mypy-boto3-s3==1.26.163
mysql==0.0.3
mysql-client==0.0.1
mysql-connector-python==8.0.11
mysqlclient==2.1.1
numpy==1.23.5
packaging==21.3
pandas==1.5.1
pathspec==0.9.0
pendulum==2.1.2
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
ply==3.11
prison==0.2.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.1
psutil==5.9.4
psycopg2-binary==2.9.5
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.19.3
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2022.7.1
pytzdata==2020.1
PyYAML==6.0
redshift-connector==2.0.911
requests==2.28.2
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.3.2
s3transfer==0.6.0
scramp==1.4.4
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.46
SQLAlchemy-JSONField==1.0.1.post0
sqlalchemy-redshift==0.8.14
SQLAlchemy-Utils==0.40.0
sqlparse==0.4.3
swagger-ui-bundle==0.0.9
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.2.0
text-unidecode==1.3
tornado==6.2
typing_extensions==4.5.0
uc-micro-py==1.0.1
unicodecsv==0.14.1
urllib3==1.26.15
vine==5.0.0
watchtower==2.0.1
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
yarl==1.8.2
zipp==3.15.0
